refactor: move http related types to http/
