+++ hals/src/http/cors.types.ts
